visitor_system/pyproject.toml
[tool.poetry]
name = "university pass system for aitu"
version = "0.1"
description = ""
authors = ["Sagat Akimbay"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.0.0"
django-allauth = "^65.7.0"
requests = "^2.32.3"
pillow = "^11.2.1"
python-dotenv = "^1.1.0"
pyopenssl = "^25.0.0"
werkzeug = "^3.1.3"
django-extensions = "^4.1"
django-select2 = "^8.4.0"
django-filter = "^25.1"
openpyxl = "^3.1.5"
celery = "^5.5.1"
redis = "^5.2.1"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.10"
pyngrok = "^7.2.5"
gunicorn = "^23.0.0"
django-pwa = "^2.0.1"
django-widget-tweaks = "^1.5.0"
faker = "^37.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
